Load all transitive dependent packages in shallow compilation (#18)
Packages may have import/export declarations, in which case the single
shallow dependency grab is not enough. Packages may also contain types which may also reference other
packages.

This idea can be expanded in the future to add necessary modules based
on diagnostics in the first pass, however reaching two layers down in
the hierarchy is generally considered a bad practice and therefore not
as common.