Make SLANG_SERVER_TESTS and env var (#23)
Having `SLANG_SERVER_TESTS` be a compiler flag means you can't use a
build for both CI and integration testing. It's also confusing when the
server silently fails to pick up the home dir conf file. Making this an
environment variable to circumvent this frustration / confusion.

Also, the python tests already thought it was an env var? See
`conftest.py`.