vaporview integration (#65)
### vaporview integration
- In Vaporview, press 'e' on a selected signal, or use the buttons to
open the corresponding filelist and go to that signal
- In Slang's Hierarchy view, press 'w' or the buttons to open the
corresponding waveform file and open in the waves
#### Button changes
- Press `cmd+f` in the hierarchy or modules view to search for an
instance in the design. Can also use the button to get to this command
(magnifying glass)
  - Select buildfile button now looks like a folder
- New 'Open buildfile' opens the current .f file. Saving this file will
refresh a compilation. In the future this will also be generated for
compilations set by top level, making it so that users can adjust the
params and defines
- "Copy Path" can be done now with either `cmd+c` or by right clicking
on a signal
####  Backend fixes that apply to both vscode and neovim:
  - Type aliases and resolved types are shown in the hierarchy view
- Macro defined symbols now show the proper locations of the macro usage
that created them