Completion: Recommend keywords on the LHS of expressions (#152)
This PR makes it so that the completion system will recommend certain
keywords on the LHS of expressions, currently: `logic, assign, wire,
reg` and snippets for `always_comb, always_latch` and `always_ff`.

I also updated the unit tests too.

This is more or less a direct-ish port of code I wrote a few years back
for my own SystemVerilog language server:
https://github.com/mattyoung101/slingshot/blob/legacy/src/main/kotlin/slingshot/completion/CompletionGenerator.kt#L66

My plan in the future is to also recommend user-defined types, like the
ones you create in a `typedef enum { ... } Type_t;`